+++ apps/ledger-live-desktop/src/renderer/families/cardano/index.ts
@@ -8,12 +8,14 @@ import AccountBodyHeader from "./AccountBodyHeader";
 import AccountSubHeader from "./AccountSubHeader";
 import sendAmountFields from "./SendAmountFields";
 import AccountBalanceSummaryFooter from "./AccountBalanceSummaryFooter";
+import accountHeaderManageActions from "./AccountHeaderManageActions";
 
 const family: LLDCoinFamily<CardanoAccount, Transaction, TransactionStatus> = {
   AccountBodyHeader,
   AccountSubHeader,
   sendAmountFields,
   AccountBalanceSummaryFooter,
+  accountHeaderManageActions,
 };
 
 export default family;