+++ apps/ledger-live-mobile/e2e/models/manager/managerPage.ts
@@ -1,3 +1,4 @@
+import { waitFor } from "detox";
 import { getElementById, openDeeplink } from "../../helpers";
 
 const baseLink = "myledger";
@@ -7,5 +8,9 @@ export default class ManagerPage {
 
   async openViaDeeplink() {
     await openDeeplink(baseLink);
+  }
+
+  async waitForManagerPageToLoad() {
+    await waitFor(this.managerTitle()).toBeVisible();
   }
 }