+++ apps/ledger-live-mobile/src/screens/Manager/Device/index.tsx
@@ -5,8 +5,7 @@ import { Trans } from "react-i18next";
 import { State, AppsDistribution, Action } from "@ledgerhq/live-common/apps/index";
 import { App, DeviceInfo, idsToLanguage } from "@ledgerhq/types-live";
 
-import { Flex, Text, Button, Divider } from "@ledgerhq/native-ui";
-import { CircledCheckSolidMedium } from "@ledgerhq/native-ui/assets/icons";
+import { Flex, Text, Button, Divider, IconsLegacy } from "@ledgerhq/native-ui";
 import styled, { useTheme } from "styled-components/native";
 import { ListAppsResult } from "@ledgerhq/live-common/apps/types";
 import { isDeviceLocalizationSupported } from "@ledgerhq/live-common/manager/localization";
@@ -135,7 +134,7 @@ const DeviceCard = ({
             </Text>
           </Flex>
           <Flex flexDirection={"row"} alignItems={"center"} mt={2} mb={3}>
-            <CircledCheckSolidMedium size={18} color={"palette.success.c40"} />
+            <IconsLegacy.CircledCheckSolidMedium size={18} color={"palette.success.c50"} />
             <Text
               variant={"body"}
               fontWeight={"medium"}