+++ libs/ledger-live-common/src/families/cardano/networks.ts
@@ -12,6 +12,7 @@ export const getNetworkParameters = (networkName: string): CardanoLikeNetworkPar
       shelleySlotDuration: 1000,
       shelleySlotsPerEpoch: 432000,
       addressPrefix: "addr",
+      poolIdPrefix: "pool",
     };
   } else if (networkName === "cardano_testnet") {
     return {
@@ -24,6 +25,7 @@ export const getNetworkParameters = (networkName: string): CardanoLikeNetworkPar
       shelleySlotDuration: 1000,
       shelleySlotsPerEpoch: 432000,
       addressPrefix: "addr_test",
+      poolIdPrefix: "pool",
     };
   }