+++ libs/ui/packages/crypto-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ledgerhq/icons-ui
 
+## 0.5.1
+
+### Patch Changes
+
+- [#4259](https://github.com/LedgerHQ/ledger-live/pull/4259) [`5696f24b93`](https://github.com/LedgerHQ/ledger-live/commit/5696f24b93151bc0ee063d1cb88cef1e2d052f9e) Thanks [@live-github-bot](https://github.com/apps/live-github-bot)! - publish ui libraries with fixed tests
+
 ## 0.5.0
 
 ### Minor Changes